requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
Django==3.0
django-cors-headers==3.4.0
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
mysqlclient==2.0.1
paystackapi==2.0.0
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
rest-condition==1.0.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
